+++ RELEASES.md
@@ -1,6 +1,8 @@
 See also GitHub releases https://github.com/lucmoreau/ProvToolbox/releases
 
 `2.2.3-SNAPSHOT`
+* prov-template: update to instantiation
+* prov-template: variable-checker: pretty printing name
 
 `2.2.2`
 * prov-template: added ptm-merging template